# internal/handler/api_handler.go
package handler

import (
	"net/http"
	"time"

	"github.com/labstack/echo/v4"

	"dynamo-modeling/internal/adapter/openapi"
	"dynamo-modeling/internal/infrastructure"
)

// APIHandler implements the OpenAPI server interface
type APIHandler struct {
	dbClient *infrastructure.DynamoDBClient
}

// NewAPIHandler creates a new API handler
func NewAPIHandler(dbClient *infrastructure.DynamoDBClient) *APIHandler {
	return &APIHandler{
		dbClient: dbClient,
	}
}

// Verify APIHandler implements ServerInterface
var _ openapi.ServerInterface = (*APIHandler)(nil)

// ListCustomers handles listing all customers
func (h *APIHandler) ListCustomers(ctx echo.Context, params openapi.ListCustomersParams) error {
	now := time.Now()

	// サンプルデータを返す（実際の実装では repository から取得）
	customers := []openapi.CustomerResponse{
		{
			Id:        "customer-1",
			Name:      "Sample Customer 1",
			Email:     "customer1@example.com",
			CreatedAt: now,
			UpdatedAt: now,
		},
		{
			Id:        "customer-2",
			Name:      "Sample Customer 2",
			Email:     "customer2@example.com",
			CreatedAt: now,
			UpdatedAt: now,
		},
	}

	return ctx.JSON(http.StatusOK, customers)
}

// CreateCustomer handles customer creation
func (h *APIHandler) CreateCustomer(ctx echo.Context) error {
	var request openapi.CustomerRequest
	if err := ctx.Bind(&request); err != nil {
		return ctx.JSON(http.StatusBadRequest, openapi.Error{
			Code:    "invalid_request",
			Message: "Invalid request body",
		})
	}

	// 現在時刻を設定
	now := time.Now()

	// レスポンスを作成
	response := openapi.CustomerResponse{
		Id:        "temp-customer-id",
		Name:      request.Name,
		Email:     request.Email,
		CreatedAt: now,
		UpdatedAt: now,
	}

	return ctx.JSON(http.StatusCreated, response)
}

// DeleteCustomer handles customer deletion
func (h *APIHandler) DeleteCustomer(ctx echo.Context, customerId string) error {
	return ctx.NoContent(http.StatusNoContent)
}

// GetCustomer handles getting a customer by ID
func (h *APIHandler) GetCustomer(ctx echo.Context, customerId string) error {
	now := time.Now()

	// サンプルデータを返す（実際の実装では repository から取得）
	response := openapi.CustomerResponse{
		Id:        customerId,
		Name:      "Sample Customer",
		Email:     "sample@example.com",
		CreatedAt: now,
		UpdatedAt: now,
	}

	return ctx.JSON(http.StatusOK, response)
}

// UpdateCustomer handles customer update
func (h *APIHandler) UpdateCustomer(ctx echo.Context, customerId string) error {
	return ctx.JSON(http.StatusOK, openapi.CustomerResponse{})
}

// GetCustomerOrders handles getting customer orders
func (h *APIHandler) GetCustomerOrders(ctx echo.Context, customerId string, params openapi.GetCustomerOrdersParams) error {
	return ctx.JSON(http.StatusOK, []openapi.OrderResponse{})
}

// ListOrders handles listing all orders
func (h *APIHandler) ListOrders(ctx echo.Context, params openapi.ListOrdersParams) error {
	return ctx.JSON(http.StatusOK, []openapi.OrderResponse{})
}

// CreateOrder handles order creation
func (h *APIHandler) CreateOrder(ctx echo.Context) error {
	return ctx.JSON(http.StatusCreated, openapi.OrderResponse{})
}

// GetOrder handles getting an order by ID
func (h *APIHandler) GetOrder(ctx echo.Context, orderId string) error {
	return ctx.JSON(http.StatusOK, openapi.OrderResponse{})
}

// UpdateOrderStatus handles order status update
func (h *APIHandler) UpdateOrderStatus(ctx echo.Context, orderId string) error {
	return ctx.JSON(http.StatusOK, openapi.OrderResponse{})
}

// ListProducts handles listing all products
func (h *APIHandler) ListProducts(ctx echo.Context, params openapi.ListProductsParams) error {
	return ctx.JSON(http.StatusOK, []openapi.ProductResponse{})
}

// CreateProduct handles product creation
func (h *APIHandler) CreateProduct(ctx echo.Context) error {
	return ctx.JSON(http.StatusCreated, openapi.ProductResponse{})
}

// DeleteProduct handles product deletion
func (h *APIHandler) DeleteProduct(ctx echo.Context, productId string) error {
	return ctx.NoContent(http.StatusNoContent)
}

// GetProduct handles getting a product by ID
func (h *APIHandler) GetProduct(ctx echo.Context, productId string) error {
	return ctx.JSON(http.StatusOK, openapi.ProductResponse{})
}

// UpdateProduct handles product update
func (h *APIHandler) UpdateProduct(ctx echo.Context, productId string) error {
	return ctx.JSON(http.StatusOK, openapi.ProductResponse{})
}